# docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_Doomed07_Bookshelf_internal_core_transport_http_response.ErrorResponse:
    properties:
      error:
        example: invalid argument
        type: string
      message:
        example: the provided argument is invalid
        type: string
    type: object
...
  internal_featurs_users_transport_http.GetUserResponse:
    properties:
      email:
        example: lol@mail.com
        type: string
      id:
        example: 1
        type: integer
      username:
        example: book_worm07
        type: string
      version:
        example: 3
        type: integer
    type: object
  internal_featurs_users_transport_http.PatchUserRequest:
    properties:
      email:
        example: Max@mail.com
        type: string
      username:
        example: booklover
        type: string
    type: object
  internal_featurs_users_transport_http.PatchedUserResponse:
    properties:
      email:
        example: lol@mail.com
        type: string
      id:
        example: 1
        type: integer
      username:
        example: book_worm07
        type: string
      version:
        example: 3
        type: integer
    type: object
...
host: 127.0.0.1:8080
info:
  contact: {}
  description: Bookshelf Application REST-API schema
  title: Bookshelf API
  version: "1.0"
paths:
...
  /users/{id}:
    delete:
      description: Delete user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: Successed to delete user
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_Doomed07_Bookshelf_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/github_com_Doomed07_Bookshelf_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_Doomed07_Bookshelf_internal_core_transport_http_response.ErrorResponse'
      summary: Delete user
      tags:
      - users
    get:
      description: Get user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: User found
          schema:
            $ref: '#/definitions/internal_featurs_users_transport_http.GetUserResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_Doomed07_Bookshelf_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/github_com_Doomed07_Bookshelf_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_Doomed07_Bookshelf_internal_core_transport_http_response.ErrorResponse'
      summary: Get user
      tags:
      - users
    patch:
      consumes:
      - application/json
      description: |-
        Partially update user fields
        ###Logic of updating fields:
        1. **If a field is provided in the request**: it will be updated with the new value.
        2. **If a field is not provided in the request**: it will remain unchanged.
        3. **If a field is provided with a null value**: it will result in a 400 error.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: PatchUser request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_featurs_users_transport_http.PatchUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successed to patch user
          schema:
            $ref: '#/definitions/internal_featurs_users_transport_http.PatchedUserResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_Doomed07_Bookshelf_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/github_com_Doomed07_Bookshelf_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_Doomed07_Bookshelf_internal_core_transport_http_response.ErrorResponse'
      summary: Patch user
      tags:
      - users
...
swagger: "2.0"